% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/csCompare.R
\name{csCompare}
\alias{csCompare}
\title{Statistically compare two CSs}
\usage{
csCompare(cs1, cs2, group = NULL, data = NULL, alternative = "two.sided",
  conf.level = 0.95, mu = 0, rscale = 0.707, descriptives = TRUE)
}
\arguments{
\item{cs1,cs2}{a numeric vector of values. If data is defined, it can refer
to either the column index or the column name of the data object.See details.}

\item{group}{column index or name that contain the group data. See details.}

\item{data}{numeric matrix or data frame that contains all data.}

\item{alternative}{a character string for the speficication of
the alternative hypothesis. Possible values: \code{"two.sided"} (default),
\code{"greater"} or \code{"less"}.}

\item{conf.level}{confidence level of the interval.}

\item{mu}{a numeric value for the mean value or mean difference}

\item{rscale}{the scale factor for the prior used in the Bayesian t.test.}

\item{descriptives}{Returns basic descriptive statistics for the dependent
variable(s).}
}
\description{
Compare two CSs within a frequentist and a Bayesian framework.
}
\details{
\code{csCompare} performs both a student t-test (using the
\code{stats::t.test} function) and a Bayesian t-test (using the
\code{BayesFactor::ttest.tstat}). If \code{cs1} and/or \code{cs2}
are or refer to multiple columns of a matrix or a data.frame, then
the row means are computed before the t-tests are performed.
In case \code{group} is \code{NULL},
paired-samples t-tests are run. In case the \code{group} is different
than \code{NULL}, then the csCompare first computes difference scores between
 the cs1 and the cs2 (i.e., cs1 - cs2).
In case the group argument is defined
but, after removal of NA's (\code{stats::na.omit}), only one group
is defined, a paired samples t-test is run.
In case of independent samples t-test, the function runs
a Welch's t-test.
}
\examples{
csCompare(cs1 = c(1, 2, 3, 1, 4), cs2 = c(10, 12, 12, 31, 13))
}
\seealso{
\code{\link[stats]{t.test}}, \code{\link[BayesFactor]{ttest.tstat}}
}

