% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csReport.R
\name{csReport}
\alias{csReport}
\title{Report results of conditioning data}
\usage{
csReport(csCompareObj = NULL, csSensitivityObj = NULL, save = FALSE,
  fileName = "report", alphalevel = 0.05)
}
\arguments{
\item{csCompareObj}{a list or data frame returned from
the \code{csCompare} function.}

\item{csSensitivityObj}{Sensitivity analysis results returned from the
\code{csSensitivity} function}

\item{save}{If \code{code} argument is set to \code{FALSE} (default), the
results are printed on the sceen. Otherwise, a '.txt' file with the
report is generated.}

\item{fileName}{The file name of the produced report. The argument is
ignored if \code{save} is set to \code{FALSE}.}

\item{alphalevel}{The alpha level to be used for determining significant
or non-significant results.}
}
\description{
Report results of data analyses run with the \code{csCompare}.
}
\examples{
tmp <- csCompare(cs1 = c(1, 2, 3, 1, 4), cs2 = c(10, 12, 12, 31, 13))
csReport(tmp)
}

