% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/csRobustnessPlot.R
\name{csRobustnessPlot}
\alias{csRobustnessPlot}
\title{Plot robusteness results}
\usage{
csRobustnessPlot(cs1, cs2, group = NULL, data = NULL,
  alternative = "two.sided", conf.level = 0.95, mu = 0,
  rscaleSens = c("medium", "wide", "ultrawide"))
}
\arguments{
\item{cs1,cs2}{a numeric vector of values. If data is defined, it can refer
to either the column index or the column name of the data object.}

\item{group}{column index or name that contain the group data. See details.}

\item{data}{numeric matrix or data frame that contains all data.}

\item{alternative}{a character string for the speficication of
the alternative hypothesis. Possible values: \code{"two.sided"} (default),
\code{"greater"} or \code{"less"}.}

\item{mu}{a numeric value for the mean value or mean difference}

\item{rscaleSens}{the scale factor for the prior used in the Bayesian t.test}
}
\description{
Perform a sensitivity analysis for the Bayes factors computed
with the \code{csCompare} results
}
\details{
This plot is influenced by the JASP way to plot robustness results.
}
\seealso{
\code{\link[condir]{csCompare}}, \code{\link[stats]{t.test}},
\code{\link[BayesFactor]{ttest.tstat}}
}

