% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csTable.R
\name{csTable}
\alias{csTable}
\title{Produce tables of \code{csCompare} results}
\usage{
csTable(csCompareObj, typeOfTable = c("latex"))
}
\arguments{
\item{csCompareObj}{a list or data frame returned from
the \code{csCompare} function}

\item{typeOfTable}{The type of table to be generated. See details}
}
\description{
It provides wrapper functions for generating tables of different
types of computer languages.
}
\details{
\code{csCompare} generates tables of different languages.
 The options are \code{latex}, and \code{markdown}.
}
\examples{
tmp <- csCompare(cs1 = c(1, 2, 3, 1, 4), cs2 = c(10, 12, 12, 31, 13))
csTable(tmp)
}

