% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundBF.R
\name{roundBF}
\alias{roundBF}
\title{Local function for determining how BF is reported}
\usage{
roundBF(bf, rscale, BF01 = TRUE)
}
\arguments{
\item{bf}{Logical. The BF.}

\item{rscale}{The rscale that was used.}

\item{BF01}{Whether the BF01 (default) should be reported or not}
}
\description{
Function used for determining the symbol, and the form, of the
BF reported.
}

