% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csReport.R
\name{csReport}
\alias{csReport}
\title{Report results of conditioning data}
\usage{
csReport(csCompareObj = NULL, csSensitivity = NULL, dir = NULL,
  fileName = "report")
}
\arguments{
\item{csCompareObj}{a list or data frame returned from
the \code{csCompare} function}

\item{dir}{The directory that the results will be saved. If set to NULL
(default), the results are saved in the current working directory.}

\item{fileName}{The file name of the produced report.}

\item{csSensitivityObj}{Sensitivity analysis results returned from the
\code{csSensitivity} function}
}
\description{
Report results of data analyses run with the \code{csCompare}.
}

