% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csRobustnessPlot.R
\name{csRobustnessPlot}
\alias{csRobustnessPlot}
\title{Plot robusteness results}
\usage{
csRobustnessPlot(cs1, cs2, group = NULL, data = NULL,
  alternative = "two.sided", conf.level = 0.95, mu = 0,
  rscaleSens = c("medium", "wide", "ultrawide"), BF01 = TRUE, ylimz = c(0,
  10), ...)
}
\arguments{
\item{cs1, cs2}{a numeric vector of values. If data is defined, it can refer
to either the column index or the column name of the data object.}

\item{group}{column index or name that contain the group data. See details.}

\item{data}{numeric matrix or data frame that contains all data.}

\item{alternative}{a character string for the speficication of
the alternative hypothesis. Possible values: \code{"two.sided"} (default),
\code{"greater"} or \code{"less"}.}

\item{conf.level}{confidence level of the interval.}

\item{mu}{a numeric value for the mean value or mean difference.}

\item{rscaleSens}{the scale factor for the prior used in the Bayesian t.test}

\item{BF01}{Should the BF01 be plotted (Default is set to TRUE). If FALSE,
the BF10 is plotted.}

\item{ylimz}{the limits of the y-axis.}

\item{...}{Further plotting arguments.}
}
\description{
Plots the results of robustness test
}
\details{
This plot is influenced by the JASP way to plot robustness results. On the
x-axis or the width of the Cauchy's Scale is plotted. On the y-axis either
BF01 is plotted (if \code{BF01} is set to TRUE) or
BF10 (if \code{BF01} is set to FALSE).
}
\examples{
csRobustnessPlot(cs1 = rnorm(10, 10, 2), cs2 = rnorm(10, 9, 3))
}
\seealso{
\code{\link[condir]{csCompare}}, #' \code{\link[condir]{csSensitivity}}
}

