% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csShine.R
\name{csShine}
\alias{csShine}
\title{Shiny app for the \code{condir} package}
\usage{
csShine()
}
\description{
Lanches a Shiny app for performing the core analyses included
in \code{condir}
}
\details{
The function can be called without any arguments (i.e., \code{csShine()}). 
For the interface, we used a \code{css} template available 
at http://getbootstrap.com.
}
\references{
Krypotos, A. M., Klugkist, I., & Engelhard, I. M. (2017). 
Bayesian hypothesis testing for human threat conditioning research: 
An introduction and the condir R package. 
European Journal of Psychotraumatology, 8.
}
