% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/csTable.R
\name{csTable}
\alias{csTable}
\title{Produce tables of \code{csCompare} results}
\usage{
csTable(csCompareObje, typeOfTable = c("latex"))
}
\arguments{
\item{csCompareObje}{a data frame returned from the \code{csCompare} object}

\item{typeOfTable}{The type of table to be generated. See details}
}
\description{
It provides wrapper functions for generating tables of different
types of computer languages.
}
\details{
\code{csCompare} generates tables of different languages.
 The options are \code{latex}, and \code{markdown}.
}

